Modifying/Cancelling listing is now possible. Any listing clicked on HomeActivity.java will allow you to either cancel a booking (if you joined as a rider) or edit/cancel the booking (if you created the listing as a driver). These changes are also accurately reflected in the DB. HomeActivity.java will now display the number of seats booked and the total amount spent for Rider bookings.
Fixed a bug in HomeActivity.java that caused Rides and Bookings to not display correctly/show duplicate bookings.
When booking a listing, if the listing does not have any more seats, it will no longer show up on the search page. Listings (or Requests) you own are also excluded.
Fixed a bug in PaymentActivity.java that allowed users to select both Credit and Debit as a payment method at the same time.
Fixed a bug in CreateRideActivity.java that caused a crash when returning back to HomeActivity.java after creating a new listing.